WIP of implementing proposal in issue #1205
